Updating rotate and flip methods to return transformed base64 image in a promise
`rotateLeft`, `rotateRight`, `flipHorizontal`, and `flipVertical` have been
updated to return a promise, which when resolved provides the transformed image
as a base64 string. This is to account for situations where autoCrop is
disabled, but a user would like access to the transformed image immediately.

When autoCrop is disabled, there is a race condition that can occur if you
invoke `crop()` immediately after transforming the image. Instead of invoking
`crop()` to retrieve the updated base64 string, it is made available from the
return value of `rotateLeft` etc.